AutomateProcesses/processShop/requirements.txt
airtable-python-wrapper==0.13.0
Flask==1.1.2
requests==2.23.0